# api/openapi.yaml
openapi: "3.1.0"
info:
  title: SoroTrail API
  description: |
    A contract event indexer for the Stellar/Soroban network. SoroTrail polls
    a Stellar RPC endpoint, stores contract events durably in Postgres, and
    serves them back through a queryable HTTP API long after the RPC has
    forgotten them.
  version: "0.1.0"
servers:
  - url: http://localhost:8080
    description: Local development
# Authentication is instance-dependent, so the default requirement lists the
# tenant credential alongside the empty requirement: a MULTI_TENANT=true
# deployment rejects an anonymous caller with 401, and a single-tenant one
# serves it. Operations that are gated regardless override this with their
# own security block.
security:
  - TenantAuth: []
  - TenantAuthHeader: []
  - {}
paths:
...
  /subscriptions:
    post:
      operationId: createSubscription
      summary: Register a new webhook subscription
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
            example:
              url: "https://example.com/webhook"
              secret: "whsec_z8eP5qL3vR2xK9yB4w"
              filters:
                contract_id: "CDLZFC3SYJYDZT7K67VZ75HPJVIEUVNIXF47ZG2FB2RMQQVU2HHGCYSC"
                type: contract
              enabled: true
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
              example:
                id: 1
                url: "https://example.com/webhook"
                filters:
                  contract_id: "CDLZFC3SYJYDZT7K67VZ75HPJVIEUVNIXF47ZG2FB2RMQQVU2HHGCYSC"
                  type: contract
                secret: "whsec_z8eP5qL3vR2xK9yB4w"
                enabled: true
                failure_count: 0
                created_at: "2026-07-21T12:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listSubscriptions
      summary: List all webhook subscriptions
      tags: [Webhooks]
      responses:
        "200":
          description: All registered subscriptions
          headers:
            X-Total-Count:
              description: >-
                The total number of subscriptions visible to the caller,
                which is the full list since this endpoint is not paginated.
              schema:
                type: integer
                minimum: 0
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
              example:
                - id: 1
                  url: "https://example.com/webhook"
                  filters:
                    contract_id: "CDLZFC3SYJYDZT7K67VZ75HPJVIEUVNIXF47ZG2FB2RMQQVU2HHGCYSC"
                  secret: "whsec_z8eP5qL3vR2xK9yB4w"
                  enabled: true
                  failure_count: 0
                  created_at: "2026-07-21T12:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /subscriptions/{id}:
    get:
      operationId: getSubscription
      summary: Get a single subscription by ID
      tags: [Webhooks]
      parameters:
        - name: id
          in: path
          required: true
          description: Subscription ID, as returned by POST /subscriptions.
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 1
      responses:
        "200":
          description: The requested subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
              example:
                id: 1
                url: "https://example.com/webhook"
                filters:
                  contract_id: "CDLZFC3SYJYDZT7K67VZ75HPJVIEUVNIXF47ZG2FB2RMQQVU2HHGCYSC"
                secret: "whsec_z8eP5qL3vR2xK9yB4w"
                enabled: true
                failure_count: 0
                created_at: "2026-07-21T12:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: updateSubscription
      summary: Update a subscription
      description: Only fields present in the request body are changed.
      tags: [Webhooks]
      parameters:
        - name: id
          in: path
          required: true
          description: Subscription ID, as returned by POST /subscriptions.
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionRequest"
            example:
              url: "https://new.example.com/webhook"
              enabled: true
      responses:
        "200":
          description: The updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
              example:
                id: 1
                url: "https://new.example.com/webhook"
                filters:
                  contract_id: "CDLZFC3SYJYDZT7K67VZ75HPJVIEUVNIXF47ZG2FB2RMQQVU2HHGCYSC"
                secret: "whsec_z8eP5qL3vR2xK9yB4w"
                enabled: true
                failure_count: 0
                created_at: "2026-07-21T12:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteSubscription
      summary: Delete a subscription and its delivery history
      tags: [Webhooks]
      parameters:
        - name: id
          in: path
          required: true
          description: Subscription ID, as returned by POST /subscriptions.
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 1
      responses:
        "204":
          description: Deleted successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  schemas:
    ErrorResponse:
      type: object
      description: Standard error envelope
      properties:
        error:
          type: string
          description: Human-readable error message
      required: [error]
...
    CreateSubscriptionRequest:
      type: object
      required: [url, secret]
      properties:
        url:
          type: string
          format: uri
          description: Callback URL
          example: "https://example.com/webhook"
        secret:
          type: string
          description: HMAC-SHA256 secret for payload verification
          example: "whsec_z8eP5qL3vR2xK9yB4w"
        filters:
          $ref: "#/components/schemas/SubscriptionFilter"
        enabled:
          type: boolean
          description: Defaults to true
    UpdateSubscriptionRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
        secret:
          type: string
        filters:
          $ref: "#/components/schemas/SubscriptionFilter"
        enabled:
          type: boolean
    SubscriptionFilter:
      type: object
      description: Filter that callbacks use to select which events to deliver. An empty filter matches every event.
      properties:
        contract_id:
          type: string
        type:
          type: string
          enum: [contract, system, diagnostic]
        topic:
          description: Exact JSON match against any topic position
        from_ledger:
          type: integer
        to_ledger:
          type: integer
    Subscription:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        filters:
          $ref: "#/components/schemas/SubscriptionFilter"
        secret:
          type: string
        enabled:
          type: boolean
        failure_count:
          type: integer
        created_at:
          type: string
          format: date-time
      required: [id, url, filters, secret, enabled, failure_count, created_at]
...
  securitySchemes:
    TenantAuth:
      type: http
      scheme: bearer
      description: |
        The tenant API key issued by POST /admin/tenants/{id}/keys, presented
        as `Authorization: Bearer st_<prefix>_<secret>`. The same key is also
        accepted in an `X-API-Key` header (see TenantAuthHeader); Bearer wins
        when both are present.

        Only enforced when the instance runs with MULTI_TENANT=true. On a
        single-tenant instance every request is served as an implicit
        wildcard principal and no credential is required, which is why most
        operations list this scheme as optional rather than required.
    TenantAuthHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        The same tenant API key as TenantAuth, presented in a header instead
        of a Bearer token.
    ManagementKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        The operator's instance-wide management key, configured through the
        API_KEY environment variable. It gates the endpoints that mutate
        ingestion or read the dead-letter queue, and it is always required on
        those endpoints — independently of MULTI_TENANT, and independently of
        the tenant credential above.

        When API_KEY is unset the server fails closed: those endpoints answer
        503 rather than running unauthenticated.
  responses:
...
    BadRequest:
      description: >-
        The request could not be parsed or violates a documented constraint —
        a malformed identifier, an out-of-range limit, an unparseable cursor,
        or a JSON body the endpoint cannot decode. The message names the
        offending parameter.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: |
        No credential was presented, or the one presented is malformed,
        unknown or revoked.

        Which credential is missing depends on the endpoint's security
        requirement: endpoints requiring ManagementKey want the operator's
        `X-API-Key`, while everything else wants the tenant key and only
        rejects when the instance runs with MULTI_TENANT=true.
      headers:
        WWW-Authenticate:
          description: >-
            The authentication challenge, per RFC 7235 §4.1 — for example
            `Bearer realm="sorotrail"`. Present on the tenant-authentication
            failures; the ManagementKey path answers without a challenge
            because it accepts no Bearer token.
          schema:
            type: string
            examples: ['Bearer realm="sorotrail"']
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: >-
        The addressed resource does not exist, or is not visible to the
        caller. Event lookups answer 404 rather than 403 for an event the
        caller may not read, so the endpoint cannot be used as an existence
        oracle. The tenant and admin endpoints also answer 404 when the
        instance runs single-tenant and those routes are therefore inert.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    TooManyRequests:
      description: |
        The caller exceeded its request quota, or asked to watch more
        contracts than its quota allows.

        Rate limiting is per client — keyed on the authenticated tenant where
        one exists and on the client IP otherwise — and is only active when
        the operator configured RATE_LIMIT_RPS or a per-tenant override. The
        probe and scrape endpoints (/health, /livez, /readyz, /metrics) are
        never rate limited, so throttling cannot blind an operator to the
        throttling itself.
      headers:
        Retry-After:
          description: >-
            Whole seconds to wait before retrying, per RFC 7231 §7.1.3.
            Always at least 1, and rounded up from the limiter's own
            wait-to-next-token estimate.
          schema:
            type: integer
            minimum: 1
            examples: [1]
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: >-
        The request failed for a reason the caller cannot act on — a store or
        RPC failure, or a panic caught by the recovery middleware. The
        message is deliberately generic; the specifics are logged against the
        request's X-Request-ID rather than returned.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...